Refactor: Task Filter
