+++ src/components/views/Tasks/Tasks.jsx
@@ -6,40 +6,24 @@ import {
   editTask,
 } from "../../../store/actions/tasksActions";
 
-import debounce from "lodash.debounce";
 import swal from "sweetalert";
 import Card from "../../Card/Card";
 import Header from "../../Header/Header";
 import CreateTask from "../../CreateTask/CreateTask";
 import SkeletonCards from "../../SkeletonCards/SkeletonCards";
+import TaskFilter from "../../TaskFilter/TaskFilter";
 
 const Tasks = () => {
   const [taskList, setTaskList] = useState([]);
   const [filteredList, setFilteredList] = useState([]);
-  const [createdBy, setCreatedBy] = useState("ALL");
-  const [titleSearch, setTitleSearch] = useState("");
 
   const dispatch = useDispatch();
   const { loading, tasks, error } = useSelector((state) => state.tasksReducer);
 
   useEffect(() => {
-    dispatch(getTasks(createdBy === "ALL" ? "" : "me"));
-  }, [createdBy]);
-
-  useEffect(() => {
     setTaskList(tasks);
     setFilteredList(tasks);
   }, [tasks]);
-
-  useEffect(() => {
-    if (titleSearch) {
-      setFilteredList(
-        taskList.filter((task) => task.title.startsWith(titleSearch))
-      );
-    } else {
-      setFilteredList(taskList);
-    }
-  }, [titleSearch]);
 
   const filterByStatus = (status) =>
     filteredList
@@ -52,24 +36,6 @@ const Tasks = () => {
           editCardStatus={handleEditStatus}
         />
       ));
-
-  const handleChangePriority = (event) => {
-    if (event.currentTarget.value === "ALL") {
-      setFilteredList(taskList);
-    } else {
-      setFilteredList(
-        taskList.filter((task) => task.importance === event.currentTarget.value)
-      );
-    }
-  };
-
-  const handleChangeCreatedBy = (event) => {
-    setCreatedBy(event.currentTarget.value);
-  };
-
-  const handleSearchByTitle = debounce((event) => {
-    setTitleSearch(event?.target?.value);
-  }, 500);
 
   const handleEditStatus = (data) => {
     dispatch(editTask(data));
@@ -117,71 +83,7 @@ const Tasks = () => {
         <CreateTask />
         <section className="w-full mx-auto mt-4 lg:col-span-2">
           <h2 className="text-lg font-semibold mb-4">My Tasks</h2>
-          {/* Task Filter */}
-          <form className="flex flex-col gap-4 items-start mb-4 lg:flex-row lg:items-center">
-            {/* Filter by Task created by */}
-            <fieldset className="flex gap-4 shrink-0">
-              <legend className="sr-only">User created Filter</legend>
-              <div className="flex items-center">
-                <input
-                  id="all-users"
-                  name="created-by"
-                  type="radio"
-                  value="ALL"
-                  className="w-4 h-4 border-gray-300 focus:ring-2 focus:ring-blue-300"
-                  checked={createdBy === "ALL"}
-                  onChange={handleChangeCreatedBy}
-                />
-                <label
-                  htmlFor="all-users"
-                  className="block ml-2 text-sm font-medium text-gray-700"
-                >
-                  All Users
-                </label>
-              </div>
-              <div className="flex items-center">
-                <input
-                  id="my-tasks"
-                  name="created-by"
-                  type="radio"
-                  value="MY-TASKS"
-                  className="w-4 h-4 border-gray-300 focus:ring-2 focus:ring-blue-300"
-                  checked={createdBy === "MY-TASKS"}
-                  onChange={handleChangeCreatedBy}
-                />
-                <label
-                  htmlFor="my-tasks"
-                  className="block ml-2 text-sm font-medium text-gray-700"
-                >
-                  My Tasks
-                </label>
-              </div>
-            </fieldset>
-
-            {/* Filter by Title */}
-            <input
-              className="block w-full py-2 px-3 text-gray-700 border rounded-2xl border-gray-300 placeholder:text-gray-700"
-              type="text"
-              name="title"
-              placeholder="Filter by Title"
-              onChange={handleSearchByTitle}
-            />
-
-            {/* Filter by Priority */}
-            <select
-              className="block w-full text-gray-700 py-2 px-3 border bg-inherit border-gray-300 rounded-2xl shadow-sm"
-              name="imṕortance"
-              onChange={handleChangePriority}
-            >
-              <option value="" disabled selected hidden>
-                Filter by Priority
-              </option>
-              <option value="ALL">All</option>
-              <option value="LOW">Low</option>
-              <option value="MEDIUM">Medium</option>
-              <option value="HIGH">High</option>
-            </select>
-          </form>
+          <TaskFilter taskList={taskList} setFilteredList={setFilteredList} />
           {/* Show Tasks */}
           {!filteredList?.length ? (
             <h3 className="text-base font-semibold mb-2">
